+++ package.json
@@ -15,6 +15,7 @@
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
     "react-native-elements": "^0.19.1",
+    "react-native-maps": "^0.22.1",
     "react-navigation": "^3.0.1"
   },
   "devDependencies": {
